requirements.txt
ahk==0.9.0
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-admin-tools==0.9.0
django-selenium==0.9.8
djangorestframework==3.11.0
freeze==3.0
idna==2.10
Jinja2==2.11.2
logger==1.4
MarkupSafe==1.1.1
mysqlclient==2.0.1
Pillow==7.2.0
pkg-resources==0.0.0
PyMySQL==0.10.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
uuid==1.30
